refactor(core): 拆分 models 大杂烩，重构核心目录并解耦引擎类型
- 移除 core/models 目录，拆分为 entities, exceptions, types 三个语义明确的子目录
- 统一重命名部分文件使其语义更清晰 (如 structs.py -> common.py, errors.py -> api_errors.py)
- 从 editor_session.py 中抽离所有数据结构与枚举到 types/editor_types.py
- 修复全局 import 路径，解决 UI 层对业务引擎的越权依赖问题
