+++ src/danmaku_sender/core/exceptions/api_errors.py
