+++ src/danmaku_sender/core/exceptions/exceptions.py
