+++ src/danmaku_sender/core/types/common.py
